Refactor: separate dir for global components
