+++ src/index.css
@@ -1,33 +1,38 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap');
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 body {
-  margin: 0;
-  font-family: 'Poppins', sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+    margin: 0;
+    font-family: 'Poppins', sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
 }
 
 h1 {
     font-family: 'Poppins', 'sans-serif';
     font-weight: 700;
 }
+
 h2,
 h3 {
     font-family: 'Poppins', 'sans-serif';
     font-weight: 600;
 }
+
 h4,
 h5,
 h6 {
     font-family: 'Poppins', 'sans-serif';
     font-weight: 500;
 }
+
 p, td {
     font-family: 'Poppins', 'sans-serif';
     font-weight: 400;
 }
+
 span,
 a,
 button,
@@ -35,8 +40,9 @@ li {
     font-family: 'Poppins', 'sans-serif';
     font-weight: 300;
 }
+
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
 
@@ -48,32 +54,4 @@ code {
     -webkit-border-radius: 10px;
     border-radius: 10px;
     background: #c0c2c5;
-}
-.back-to-top {
-    display: inline-block;
-    background: linear-gradient(to left, #f12711, #f5af19);
-    width: 50px;
-    height: 50px;
-    text-align: center;
-    border-radius: 7px;
-    position: fixed;
-    bottom: 80px;
-    right: 25px;
-    transition: all 1.3s;
-    visibility: visible;
-    z-index: 300000;
-    color: #FFF;
-    font-size: 30px;
-}
-.back-to-top:hover {
-    cursor: pointer;
-}
-.fade-out{
-    opacity:0;
-    transition: all 1.3s;
-
-}
-.fade-in {
-    opacity: 1;
-    transition: all 1.3s;
 }
